+++ ChangeLog.md
@@ -41,6 +41,10 @@
     problems, add Dirichlet boundary conditions, update discrete
     differential calculus examples (Jacques-Olivier
     Lachaud,[#1643](https://github.com/DGtal-team/DGtal/pull/1643))
+  - Updates to PolygonalCalculus: adding vector field operators (mainly covariant 
+    gradient and covariant projection as well as Connection-Laplacian). Also adding two
+    more examples: harmonic parametrization and vectors in heat method. (Baptiste GENEST, David
+    Coeurjolly,  [#1646](https://github.com/DGtal-team/DGtal/pull/1646))
 
 - *Topology*
   - New helper methods to retrieve the interior/exterior voxel of a given